requirements.txt
requests
beautifulsoup4
chromadb
google-generativeai